Added more options to start.sh
The VMs supplied by modern.ie are currently 32b systems.  I was
having issues running these with the packaged version of
qemu-system-x86_64 on my distro (Ubuntu 16.04) and wanted to be
able to have a flag to use qemu-system-i386.  In doing so, I
also added a few more options that made start.sh more flexible.

--32bqemu
        use qemu-system-i386

-m
        RAM for system in MB

--pcnet
        Use pcnet network adaptor.

--image imagename
        use imagename for qemu-system

--help/-?
        A (probably excessively) terse help output.

This patch does admittedly break the current "just pass an image
name" on the command line.  I don't know how to reconcile that.